requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.133.0
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
oauthlib==3.3.1
packaging==26.0
postgrest==2.28.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.28.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
